reorganized component files to allow for unit tests
